+++ .env
@@ -1,0 +1,2 @@
+REACT_APP_TITLE="React Redux Json Placeholder"
+REACT_APP_TEST="It works!!!"
