+++ src/containers/components/link.js
@@ -4,8 +4,19 @@ import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
 import PropTypes from "prop-types";
+import makeStyles from "@material-ui/core/styles/makeStyles";
 
+const useStyles = makeStyles((theme) => ({
+  link: {
+    padding: "20px 5px",
+  },
+  icon: {
+    color: "blue",
+    background: "blue",
+  },
+}));
 function ListItemLink(props) {
+  const classes = useStyles();
   const { icon, primary, to } = props;
 
   const CustomLink = React.useMemo(
@@ -19,8 +30,8 @@ function ListItemLink(props) {
   return (
     <li>
       <ListItem button component={CustomLink}>
-        <ListItemIcon>{icon}</ListItemIcon>
-        <ListItemText primary={primary} />
+        <ListItemIcon className={classes.link}>{icon}</ListItemIcon>
+        <ListItemText className={classes.link} primary={primary} />
       </ListItem>
     </li>
   );
