+++ src/views/albums/index.jsx
@@ -1,0 +1,3 @@
+import Albums from "./albums";
+
+export default Albums;
