+++ src/views/comments/index.jsx
@@ -1,0 +1,3 @@
+import Comments from "./comments";
+
+export default Comments;
