+++ src/views/photos/index.jsx
@@ -1,0 +1,3 @@
+import PhotosView from "./photos";
+
+export default PhotosView;
