+++ src/views/posts/index.jsx
@@ -1,0 +1,3 @@
+import PostsView from "./posts";
+
+export default PostsView;
