+++ src/views/todos/index.jsx
@@ -1,0 +1,3 @@
+import TodosView from "./todos";
+
+export default TodosView;
