+++ src/views/users/index.jsx
@@ -1,0 +1,3 @@
+import UsersView from "./users";
+
+export default UsersView;
